lesson_22/requirements.txt
Flask>=3.0.2
requests>=2.31.0
python-dotenv>=1.0.0